requirements.txt
fastapi==0.115.6
langchain==0.3.12
langchain_chroma==0.1.4
langchain_community==0.3.12
langchain_core==0.3.25
langchain_openai==0.2.12
langchain_text_splitters==0.3.3
pydantic==2.10.3
python-dotenv==1.0.1
chromadb==0.5.20
jinja2
elevenlabs
google-cloud-speech
python-multipart
ffmpeg-python  # Python wrapper for FFmpeg